Split parser part of search.rb into multiple pieces
